codigo modularizado y documentado
documentado para los pibes
